pyproject.toml
[project]
name = "grader"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.14"
dependencies = [
    "dataclasses-json>=0.6.7",
    "gitpython>=3.1.45",
    "pyyaml==6.0.3",
    "requests>=2.32.5",
    "submitit>=1.5.4",
]
